resize problems with new window
