+++ Engine/imgui.ini
@@ -10,8 +10,8 @@ Collapsed=0
 DockId=0x00000004,1
 
 [Window][Hierarchy]
-Pos=772,19
-Size=293,791
+Pos=555,19
+Size=510,791
 Collapsed=0
 DockId=0x00000003,0
 
@@ -40,8 +40,8 @@ Collapsed=0
 DockId=0x00000004,2
 
 [Window][Console]
-Pos=0,417
-Size=770,393
+Pos=0,343
+Size=553,467
 Collapsed=0
 DockId=0x00000007,0
 
@@ -62,21 +62,21 @@ Collapsed=0
 
 [Window][Scene]
 Pos=0,19
-Size=770,396
+Size=553,322
 Collapsed=0
 DockId=0x00000002,0
 
 [Window][Preview Shapes FBO]
-Pos=577,201
-Size=685,411
+Pos=471,119
+Size=611,539
 Collapsed=0
 
 [Docking][Data]
-DockSpace     ID=0x09EF459F Window=0x9A404470 Pos=240,214 Size=1440,791 Split=X Selected=0x18B8C0DE
-  DockNode    ID=0x00000001 Parent=0x09EF459F SizeRef=770,791 Split=Y Selected=0x18B8C0DE
-    DockNode  ID=0x00000002 Parent=0x00000001 SizeRef=515,396 CentralNode=1 Selected=0x18B8C0DE
-    DockNode  ID=0x00000007 Parent=0x00000001 SizeRef=515,393 Selected=0xF9BEF62A
-  DockNode    ID=0x00000006 Parent=0x09EF459F SizeRef=668,791 Split=X Selected=0x169CEE22
-    DockNode  ID=0x00000003 Parent=0x00000006 SizeRef=293,791 Selected=0x788BAA0D
+DockSpace     ID=0x09EF459F Window=0x9A404470 Pos=120,139 Size=1440,791 Split=X Selected=0x18B8C0DE
+  DockNode    ID=0x00000001 Parent=0x09EF459F SizeRef=553,791 Split=Y Selected=0x18B8C0DE
+    DockNode  ID=0x00000002 Parent=0x00000001 SizeRef=515,322 CentralNode=1 Selected=0x18B8C0DE
+    DockNode  ID=0x00000007 Parent=0x00000001 SizeRef=515,467 Selected=0xF9BEF62A
+  DockNode    ID=0x00000006 Parent=0x09EF459F SizeRef=885,791 Split=X Selected=0x169CEE22
+    DockNode  ID=0x00000003 Parent=0x00000006 SizeRef=510,791 Selected=0x788BAA0D
     DockNode  ID=0x00000004 Parent=0x00000006 SizeRef=373,791 Selected=0x74350DE7
